+++ plugins/features/plugin-photo-swipe/src/client/utils/registerPhotoSwipe.ts
@@ -3,25 +3,19 @@ import type PhotoSwipe from 'photoswipe'
 import type { SlideData } from 'photoswipe'
 import type { PhotoSwipeOptions } from '../helpers/index.js'
 import { LOADING_ICON } from './icon.js'
-import { getImageUrlInfo } from './images.js'
+import { getImageElementInfo } from './images.js'
 import { setupPhotoSwipe } from './setupPhotoSwipe.js'
 import type { PhotoSwipeBehaviorOptions } from './typings.js'
 
-export interface PhotoSwipeState {
-  open: (index: number) => void
-  close: () => void
-  destroy: () => void
-}
-
-export const createPhotoSwipe = async (
-  images: string[],
+export const registerPhotoSwipe = async (
+  images: HTMLImageElement[],
   {
     scrollToClose = true,
     download = true,
     fullscreen = true,
     ...photoSwipeOptions
   }: PhotoSwipeBehaviorOptions & PhotoSwipeOptions,
-): Promise<PhotoSwipeState> => {
+): Promise<() => void> => {
   const { default: PhotoSwipe } = await import(
     /* webpackChunkName: "photo-swipe" */ 'photoswipe'
   )
@@ -29,24 +23,13 @@ export const createPhotoSwipe = async (
 
   const dataSource = images.map<SlideData>((image) => ({
     html: LOADING_ICON,
-    msrc: image,
+    element: image,
+    msrc: image.src,
   }))
 
   images.forEach((image, index) => {
-    void getImageUrlInfo(image).then((data) => {
-      dataSource.splice(index, 1, data)
-      currentPhotoSwipe?.refreshSlideContent(index)
-    })
-  })
-
-  const destroy = useEventListener('wheel', () => {
-    currentPhotoSwipe?.close()
-  })
-
-  return {
-    open: (index: number): void => {
-      currentPhotoSwipe?.close()
-
+    const handler = (): void => {
+      currentPhotoSwipe?.destroy()
       currentPhotoSwipe = new PhotoSwipe({
         preloaderDelay: 0,
         showHideAnimationType: 'zoom',
@@ -60,12 +43,34 @@ export const createPhotoSwipe = async (
 
       setupPhotoSwipe(currentPhotoSwipe, { download, fullscreen })
 
-      currentPhotoSwipe.addFilter('placeholderSrc', () => images[index])
+      currentPhotoSwipe.addFilter('thumbEl', () => image)
+      currentPhotoSwipe.addFilter('placeholderSrc', () => image.src)
       currentPhotoSwipe.init()
-    },
-    close: (): void => {
-      currentPhotoSwipe?.close()
-    },
-    destroy,
-  }
+    }
+
+    if (!image.getAttribute('photo-swipe')) {
+      image.style.cursor = 'zoom-in'
+      image.addEventListener('click', () => {
+        handler()
+      })
+      image.addEventListener('keypress', ({ key }) => {
+        if (key === 'Enter') handler()
+      })
+      // avoid registering multiple times
+      image.setAttribute('photo-swipe', '')
+    }
+
+    void getImageElementInfo(image).then((data) => {
+      dataSource.splice(index, 1, data)
+      currentPhotoSwipe?.refreshSlideContent(index)
+    })
+  })
+
+  return scrollToClose
+    ? useEventListener('wheel', () => {
+        currentPhotoSwipe?.close()
+      })
+    : (): void => {
+        // do nothing
+      }
 }
